+++ client/landing/stepper/declarative-flow/internals/steps-repository/ai-site-prompt/index.tsx
@@ -3,13 +3,12 @@ import styled from '@emotion/styled';
 import { TextareaControl } from '@wordpress/components';
 import { useSelect } from '@wordpress/data';
 import { useI18n } from '@wordpress/react-i18n';
-import { FormEvent, useState } from 'react';
-import { useSearchParams } from 'react-router-dom';
-import wpcomRequest from 'wpcom-proxy-request';
+import { FormEvent } from 'react';
 import FormattedHeader from 'calypso/components/formatted-header';
 import { LoadingEllipsis } from 'calypso/components/loading-ellipsis';
 import { ONBOARD_STORE } from 'calypso/landing/stepper/stores';
 import { recordTracksEvent } from 'calypso/lib/analytics/tracks';
+import useAIAssembler from '../pattern-assembler/hooks/use-ai-assembler';
 import type { Step } from '../../types';
 import type { OnboardSelect } from '@automattic/data-stores';
 
@@ -43,9 +42,7 @@ const AISitePrompt: Step = function ( props ) {
 		[]
 	);
 
-	const [ searchParams, setSearchParams ] = useSearchParams(); // eslint-disable-line @typescript-eslint/no-unused-vars
-	const [ loading, setLoading ] = useState( false );
-	const [ prompt, setPrompt ] = useState( searchParams.get( 'ai_description' ) || '' );
+	const [ callAIAssembler, setPrompt, prompt, loading ] = useAIAssembler();
 
 	const stepProgress = useSelect(
 		( select ) => ( select( ONBOARD_STORE ) as OnboardSelect ).getStepProgress(),
@@ -54,66 +51,9 @@ const AISitePrompt: Step = function ( props ) {
 
 	const onSubmit = async ( event: FormEvent ) => {
 		event.preventDefault();
-		setSearchParams( ( currentSearchParams ) => {
-			currentSearchParams.set( 'ai_description', prompt );
-			return currentSearchParams;
-		} );
-		setLoading( true );
-		wpcomRequest( {
-			path: '/pattern-assembler/ai/latest/generate',
-			method: 'POST',
-			apiNamespace: 'wpcom/v2',
-			body: {
-				description: prompt,
-			},
-		} )
-			.then( ( response: any ) => {
-				console.log( 'Patterns AI response', response ); /* eslint-disable-line no-console */
-				// This actually passes the patterns to the pattern assembler.
-				setSearchParams(
-					( currentSearchParams ) => {
-						currentSearchParams.set( 'header_pattern_id', response.header_pattern );
-						currentSearchParams.set( 'footer_pattern_id', response.footer_pattern );
-						currentSearchParams.set( 'pattern_ids', response.pages[ 0 ].patterns.join( ',' ) );
-						// These 2 params were introduced in the V5 of the AI endpoint.
-						// TODO: Remove the checks once we settle on the endpoint.
-						if ( response?.site?.site_title ) {
-							currentSearchParams.set( 'site_title', response.site.site_title );
-						}
-						if ( response?.site?.site_tagline ) {
-							currentSearchParams.set( 'site_tagline', response.site.site_tagline );
-						}
-
-						// This was introduced in the V5 of the AI endpoint.
-						const pageSlugs = response.pages.map( ( page: any ) => page?.slug ).filter( Boolean );
-						if ( pageSlugs.length ) {
-							currentSearchParams.set( 'page_slugs', pageSlugs.join( ',' ) );
-						}
-
-						if ( response.style ) {
-							currentSearchParams.set( 'color_variation_title', response.style );
-						}
-
-						if ( response.font ) {
-							currentSearchParams.set( 'font_variation_title', response.font );
-						}
-
-						// So that we close the drawer with patterns when moving to the assembler:
-						currentSearchParams.set( 'screen', 'main' );
-						currentSearchParams.delete( 'screen_parameter' );
-
-						return currentSearchParams;
-					},
-					{ replace: true }
-				);
-				setLoading( false );
-				submit?.();
-			} )
-			.catch( ( error ) => {
-				console.error( 'big sky error', error ); /* eslint-disable-line no-console */
-				setLoading( false );
-				goNext?.();
-			} );
+		callAIAssembler()
+			?.then( () => submit?.() )
+			?.catch( () => goNext?.() );
 	};
 
 	function getContent() {
